README.md added, some TODOs made.

+++ install.sh
@@ -2,17 +2,20 @@
 
 echo 'Installing dotfiles...'
 
-apt-get install  git-core vim i3 feh tmux lxappearance gtk-chtheme qt4-qtconfig -y
+apt-get install  git-core vim i3 feh tmux mdp -y # lxappearance gtk-chtheme qt4-qtconfig -y
 
 
 for file in $(ls -A $PWD) # `ls -A` can be used
 do
 	if [[ "$file" != .git* ]]; then 	
 		if [[ "$file" != *install* ]]; then
+            if [[ -L ~/$file ]]; then 
+                echo "Link for $file already exist!"
+            else
 
-			echo "Making link for $file"
-			ln -s $PWD/$file ~/$file 2>/dev/null
-
+                echo "Making link for $file"
+                ln -s $PWD/$file ~/$file 2>/dev/null
+            fi
 		fi
 	fi
 
@@ -20,18 +23,38 @@ done
 
 
 if [[ -e ~/.profile ]]; then # For Debian based distros
-
-    echo "alias tmux='tmux -2'" >>  ~/.bashrc
-    echo "[[ -r ~/.bashrc ]] && . ~/.bashrc" >>  ~/.profile
+    
+    if ! grep -q META_COMMENT_1 ~/.profile; then
+        
+        echo "#META_COMMENT_1" >> ~/.profile
+        echo "alias tmux='tmux -2'" >>  ~/.bashrc
+        echo "export TERM=xterm-256color" >>  ~/.bashrc
+        echo "[[ -r ~/.bashrc ]] && . ~/.bashrc" >>  ~/.profile
+        
+    fi
 fi 
 
 if [[ -e ~/.bash_profile ]]; then # For other distros 
 
-    echo "alias tmux='tmux -2'" >>  ~/.bashrc
-    echo "[[ -r ~/.bashrc ]] && . ~/.bashrc" >>  ~/.bash_profile
-
+    if ! grep -q META_COMMENT_1 ~/.bash_progile; then
+     echo "#META_COMMENT_1" >> ~/.profile
+        echo "alias tmux='tmux -2'" >>  ~/.bashrc
+        echo "export TERM=xterm-256color" >>  ~/.bashrc
+        echo "[[ -r ~/.bashrc ]] && . ~/.bashrc" >>  ~/.bash_profile
+    fi
  fi   
 
+ 
 
+if [[ ! -d ~/.vimbckp/ ]]; then 
+    mkdir ~/.vimbckp/
+    mkdir ~/.vimbckp/.backup ~/.vimbckp/.swp ~/.vimbckp/.undo 
+fi
 
+cd $HOME
+mkdir tmp
 
+# Write here custom instalation and compilations
+
+cd $HOME
+rm -rf tmp
